requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==1.11.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
dataclasses-json==0.6.1
Deprecated==1.2.14
dlib==19.24.6
emoji==2.8.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.104.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.329
langdetect==1.0.9
langsmith==0.0.56
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.28.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.1
pandas==2.1.3
passlib==1.7.4
Pillow==10.0.1
posthog==3.0.2
protobuf==4.21.12
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.1.1
PyPika==0.48.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyserial==3.5
pystime==0.0.1.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.20
pytime==0.2.3
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.15.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spotify==0.10.2
spotipy==2.23.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.10.28
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.15.0
XlsxWriter==3.2.2
yarl==1.9.2
zipp==3.17.0
